refactor(webhook): extract delivery into webhook_delivery mixin
Behavior-preserving verbatim extraction of _direct_deliver,
_deliver_github_comment, and _deliver_cross_platform from
gateway/platforms/webhook.py into gateway/platforms/webhook_delivery.py
as WebhookDeliveryMixin, composed into WebhookAdapter MRO.

Preserve the monkeypatch seam: keep 'import subprocess' re-exported
through webhook.py so existing tests patching
gateway.platforms.webhook.subprocess.run continue to resolve.

Also make the profile admission seam test MRO-robust (assert membership
+ method identity instead of an exact prefix) so adding further mixins
does not break the seam contract.
